Split the local storage service and worker bridge into single-purpose utils
