+++ packages/twenty-front-component-renderer/src/host/constants/FrontComponentLocalStorageDatabaseName.ts
@@ -1,0 +1,2 @@
+export const FRONT_COMPONENT_LOCAL_STORAGE_DATABASE_NAME =
+  'front-component-local-storage';
