+++ packages/twenty-front-component-renderer/src/host/constants/FrontComponentLocalStorageDatabaseVersion.ts
@@ -1,0 +1,1 @@
+export const FRONT_COMPONENT_LOCAL_STORAGE_DATABASE_VERSION = 1;
