+++ packages/twenty-front-component-renderer/src/host/constants/FrontComponentLocalStorageEntriesStoreName.ts
@@ -1,0 +1,1 @@
+export const FRONT_COMPONENT_LOCAL_STORAGE_ENTRIES_STORE_NAME = 'entries';
