+++ packages/twenty-front-component-renderer/src/host/utils/buildFrontComponentLocalStorageNamespace.ts
@@ -1,0 +1,6 @@
+import { type FrontComponentLocalStorageNamespace } from '@/types/FrontComponentLocalStorageNamespace';
+
+export const buildFrontComponentLocalStorageNamespace = ({
+  applicationId,
+  userId,
+}: FrontComponentLocalStorageNamespace): string => `${applicationId}:${userId}`;
