+++ packages/twenty-front-component-renderer/src/host/utils/buildFrontComponentLocalStorageNamespaceRange.ts
@@ -1,0 +1,3 @@
+export const buildFrontComponentLocalStorageNamespaceRange = (
+  namespace: string,
+): IDBKeyRange => IDBKeyRange.bound([namespace], [namespace, []]);
