+++ packages/twenty-front-component-renderer/src/remote/worker/utils/toQuotaExceededDomException.ts
@@ -1,0 +1,6 @@
+import { FrontComponentStorageError } from '@/utils/FrontComponentStorageError';
+
+export const toQuotaExceededDomException = (error: unknown): unknown =>
+  error instanceof FrontComponentStorageError
+    ? new DOMException(error.message, 'QuotaExceededError')
+    : error;
