+++ packages/twenty-front-component-renderer/src/types/FrontComponentLocalStorageEntry.ts
@@ -1,0 +1,7 @@
+export type FrontComponentLocalStorageEntry = {
+  namespace: string;
+  key: string;
+  value: string;
+  size: number;
+  updatedAt: number;
+};
