+++ packages/twenty-front-component-renderer/src/types/FrontComponentLocalStoragePersistQueue.ts
@@ -1,0 +1,10 @@
+import { type FrontComponentHostCommunicationApiStore } from '@/types/FrontComponentHostCommunicationApiStore';
+
+export type FrontComponentLocalStoragePersistQueue = {
+  schedule: <TResult>(
+    runPersist: (
+      hostCommunicationApi: FrontComponentHostCommunicationApiStore,
+    ) => Promise<TResult> | undefined,
+  ) => Promise<TResult>;
+  flush: () => void;
+};
